+++ example/flex_test_host/lib/main.dart
@@ -1,5 +1,12 @@
 import 'package:flutter/widgets.dart';
 
+class FlexTestHostApp extends StatelessWidget {
+  const FlexTestHostApp({super.key});
+
+  @override
+  Widget build(BuildContext context) => const SizedBox.shrink();
+}
+
 void main() {
-  runApp(const SizedBox.shrink());
+  runApp(const FlexTestHostApp());
 }