+++ example/integration_test/coreml_model_sweep_test.dart
@@ -12,10 +12,9 @@
 
 import 'dart:io';
 import 'dart:math';
-import 'dart:typed_data';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import 'package:flutter_litert/flutter_litert.dart';
+import 'package:flutter_litert/native.dart';
 
 const int iterations = 30;
 const int warmup = 8;