+++ example/integration_test/interpreter_delegate_fallback_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/services.dart' show rootBundle;
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import 'package:flutter_litert/flutter_litert.dart';
+import 'package:flutter_litert/native.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();