+++ example/integration_test/xnnpack_delegate_test.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import 'package:flutter_litert/flutter_litert.dart';
+import 'package:flutter_litert/native.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();